[Keyboard] Add Quefrency Rev. 5 (#17503)
