+++ keyboards/keebio/quefrency/keymaps/default60/config.h
@@ -21,4 +21,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 #pragma once
 
+/* Use I2C or Serial, not both */
+
+#define USE_SERIAL
 // #define USE_I2C
