+++ keyboards/keebio/quefrency/quefrency.h
@@ -14,3 +14,6 @@
 #ifdef KEYBOARD_keebio_quefrency_rev4
     #include "rev4.h"
 #endif
+#ifdef KEYBOARD_keebio_quefrency_rev5
+    #include "rev5.h"
+#endif
