+++ keyboards/keebio/quefrency/rev5/info.json
@@ -1,5 +1,5 @@
 {
-    "keyboard_name": "Quefrency rev4",
+    "keyboard_name": "Quefrency Rev. 5",
     "url": "https://keeb.io",
     "maintainer": "nooges",
     "layout_aliases": {
