+++ keyboards/keebio/quefrency/rev5/readme.md
@@ -1,0 +1,18 @@
+# Quefrency Rev. 5
+
+Quefrency Rev. 5 features Kailh hotswap sockets with a fixed layout. The left macro portion and right column can be broken off, and split backspace or 2u backspace can be used.
+
+* Keyboard Maintainer: [Bakingpy/nooges](https://github.com/nooges)
+* Hardware Availability: [Keebio](https://keeb.io/)
+
+Make example for this keyboard (after setting up your build environment):
+
+    make keebio/quefrency/rev5:default
+
+Example of flashing this keyboard:
+
+    make keebio/quefrency/rev5:default:flash
+
+See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
+
+A build guide for this keyboard can be found here: [Keebio Build Guides](https://docs.keeb.io)
