+++ keyboards/keebio/quefrency/rev5/rules.mk
@@ -1,0 +1,4 @@
+BOOTLOADER = atmel-dfu
+BACKLIGHT_ENABLE = no
+RGBLIGHT_ENABLE = yes
+ENCODER_ENABLE = yes
